requirements.txt
mysql-connector-python
string
secrets
sqlite3